requirements.txt
requests>=2.31.0
backports.zoneinfo; python_version<"3.9"
python-dotenv>=1.0.1
boto3>=1.34.0
